Fixed a problem with mapping field magnitude of flow.  Corrected a couple of compiler warnings.

+++ lib/params/Colormap.h
@@ -67,7 +67,7 @@ public:
   float maxValue() const;      // Data Coordinates
   void  maxValue(float value); // Data Coordinates
 
-  int numControlPoints()                { return _controlPoints.size(); }
+  int numControlPoints()                { return (int)_controlPoints.size(); }
 
   Color controlPointColor(int index);
   void  controlPointColor(int index, Color color);
